test-fix: move constants in the separate file and move fixture for product in the product tests file.
