+++ backend/tests/unit/constants.py
@@ -1,0 +1,15 @@
+"""The file with constats for the tests"""
+
+from uuid import UUID
+
+# Roles
+USER_ROLE = 1
+ADMIN_ROLE = 2
+NO_EXISTING_ROLE = 100
+
+# The data for product
+TEST_PRODUCT_UUID = UUID("a5e7d95f-3f06-4a31-8079-4f9c937f4775")
+TEST_PRODUCT_NAME = "Cute product"
+TEST_PRODUCT_PRICE = 100.50
+TEST_RPODUCT_BRAND_ID = 1
+TEST_PRODUCT_DISCOUNT = 0.2
