+++ backend/tests/unit/domain/entities/conftest.py
@@ -5,22 +5,10 @@ from contextlib import nullcontext as dont_raise
 from uuid import UUID
 
 from pytest import fixture
+from tests.unit.constants import ADMIN_ROLE, NO_EXISTING_ROLE, USER_ROLE
 
-from domain.entities.product import Product
+from domain.entities.products import Product, Wine
 from domain.entities.user import UserBase
-
-# Roles
-USER_ROLE = 1
-ADMIN_ROLE = 2
-NO_EXISTING_ROLE = 100
-
-
-# The data for product
-TEST_PRODUCT_UUID = UUID("a5e7d95f-3f06-4a31-8079-4f9c937f4775")
-TEST_PRODUCT_NAME = "Cute product"
-TEST_PRODUCT_PRICE = 100.50
-TEST_RPODUCT_BRAND_ID = 1
-TEST_PRODUCT_DISCOUNT = 0.2
 
 
 # the user_base data for testing
@@ -49,21 +37,3 @@ def user_base(request) -> dict[str, UserBase | AbstractAsyncContextManager]:
         "user": UserBase(user_id=user_id, role_id=role_id, login=login),
         "expectation": expectation,
     }
-
-
-# ================
-# Product fixtures
-
-
-@fixture()
-def product_base():
-    return Product(
-        product_uuid=TEST_PRODUCT_UUID,
-        name=TEST_PRODUCT_NAME,
-        price=TEST_PRODUCT_PRICE,
-        brand_id=TEST_RPODUCT_BRAND_ID,
-        discount=TEST_PRODUCT_DISCOUNT,
-    )
-
-
-# ================
